backend/requirements.txt
# Core Framework
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.0

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
PyPDF2==3.0.1
pdf2image==1.16.3

# OCR (CRITICAL)
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# AI/ML
google-generativeai==0.3.0
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2

# Report Generation
reportlab==4.0.4

# Utilities
flask-cors==4.0.0
